requirements.txt
requests
python-dotenv
openai
fastapi
uvicorn